Move meetings server to meetings package
